Use inner class instead seperate domain class
