--- app/templates/src/main/java/package/domain/_KeyAndPassword.java
@@ -1,25 +1,0 @@
-package <%=packageName%>.domain;
-
-public class KeyAndPassword {
-
-    private String key;
-    private String newPassword;
-
-    public KeyAndPassword() {}
-
-    public String getKey() {
-        return key;
-    }
-
-    public void setKey(String key) {
-        this.key = key;
-    }
-
-    public String getNewPassword() {
-        return newPassword;
-    }
-
-    public void setNewPassword(String newPassword) {
-        this.newPassword = newPassword;
-    }
-}
